+++ app.js
@@ -1,4 +1,5 @@
 const requireEnv = require('require-environment-variables');
+
 requireEnv(['ETL_NAME']);
 const etl = require('./app/lib/etl');
 const dataService = require('./app/lib/azureDataService');