+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.3.0 / 2018-04-12
 ==================
 - Add GSD ID to output data
+- Map telephone number to `contacts.telephoneNumber`
 
 0.2.0 / 2018-03-28
 ==================