+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ==================
 - Add GSD ID to output data
 - Map telephone number to `contacts.telephoneNumber`
+- Map service type to `serviceType`
 
 0.2.0 / 2018-03-28
 ==================