+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Map telephone number to `contacts.telephoneNumber`
 - Map service type to `serviceType`
 - Map opening times to `openingTimes.description`
+- Update npm dependencies
 
 0.2.0 / 2018-03-28
 ==================